// frontend/types.ts
export interface Athlete {
  id: string;
  firstName: string;
  lastName: string;
  name: string; // firstName + lastName
  birthDate: string;
  cpf: string;
  address: string;
  email: string;
  phone: string;
  category: string;
  avatarUrl?: string;
  status: 'Active' | 'Blocked';
  recentLoad?: number; 
  attendance?: boolean;
  fatigueScore?: number; 
  bodyWeight?: number;
}